[SCHEMA] Add remote_uid_count bookkeeping to FolderSync.
For better progress reporting, we need to keep a local record of how
many messages are on the remote, even if our notion will get out of date
quickly. This way we can bootstrap up a rough progress report without
having to be able to connect to the remote account backend.